perf(header): isolate usePathname() so navigation stops re-rendering the header
Header.tsx is 440 LOC and 'use client'. It used to call usePathname() at
the top of the component, which means every single <Link> navigation
re-rendered the whole header — all six Heroicons, the cart store
subscription, the categories dropdown state, everything — just to
recompute two small pathname-dependent bits:

  1. Whether to show the PWA install CTA (hidden on /admin, /vendor,
     /checkout to avoid interrupting work/buy flows).
  2. A subtle background highlight on the /login link when the user is
     already on /login.

Push those two reads into a new, tiny HeaderPathnameParts.tsx client
module that exports <InstallCtaGate /> and <LoginLink />. The Header
body no longer subscribes to pathname at all, so a navigation from
/productos to /productores only re-renders the two small components
that actually care about the path — not the 440-line parent.

Behaviour is unchanged:
  - Same CSS classes on the login link (regex-exact match on /login).
  - Same denylist prefixes on the install CTA gate.
  - Same render order (InstallCtaGate takes the same slot as the old
    {canShowInstallCta && <InstallButton />} expression).

Part of Phase B → P0-6 from docs/audits/runtime-performance-audit.md.
